+++ examples/ubldc-demo/ubldc-demo.py
@@ -26,7 +26,7 @@ load_dotenv()
 
 
 async def main():
-    ubldc.create_depth_cache(markets=markets)
+    ubldc.create_depthcache(markets=markets)
     while ubldc.is_stop_request() is False:
         add_string = f"---------------------------------------------------------------------------------------------"
         for market in markets: